test(network): add unit tests for network capability and move to dedicated directory
